Move protocol code to protocol.c
